Fix warnings that implicit conversion to float changes the value.
There are three warnings about an implicit conversion from into to
float causing the resulting float to be one higher that the original
integer.  Fix two of these by using a static cast, and the third by
using MAXFLOAT instead of INT32_MAX.

+++ mythtv/libs/libmyth/audio/audiooutpututil.cpp
@@ -213,7 +213,8 @@ char *AudioOutputUtil::GeneratePinkFrames(char *frames, int channels,
             if (chn==channel)
             {
                 /* Don't use MAX volume */
-                double res = generate_pink_noise_sample(&pink) * 0x03fffffff;
+                double res = generate_pink_noise_sample(&pink) *
+                    static_cast<float>(0x03fffffff);
                 int32_t ires = res;
                 if (bits == 16)
                     *samp16++ = LE_SHORT(ires >> 16);
